# src/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: GUI-LOP API
  description: |
    ## Generative UI & Human-in-the-Loop Orchestration Platform API

    The GUI-LOP API provides comprehensive endpoints for authentication, workflow management,
    real-time communication, and system monitoring. This API supports:

    - JWT-based authentication with refresh tokens
    - Workflow creation, execution, and management
    - Real-time WebSocket communication
    - Rate limiting and security features
    - Comprehensive error handling
    - API versioning and backward compatibility

    ### Authentication
    The API uses JWT (JSON Web Tokens) for authentication. Include the access token
    in the Authorization header:

    ```
    Authorization: Bearer <access_token>
    ```

    ### Rate Limiting
    API endpoints are rate-limited to prevent abuse. Rate limit information is
    included in response headers:

    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Time when rate limit resets (Unix timestamp)

    ### Error Responses
    All error responses follow a consistent format:

    ```json
    {
      "success": false,
      "message": "Error description",
      "code": "ERROR_CODE",
      "details": {...},
      "timestamp": "2024-01-01T00:00:00.000Z"
    }
    ```

    ### Pagination
    List endpoints support pagination using:
    - `page`: Page number (default: 1)
...

...
    ```json
    {
...
      }
    }
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /workflows:
    get:
      tags: [Workflows]
      summary: Get user workflows
      description: |
        Retrieve a paginated list of workflows belonging to the current user.
        Supports filtering by status, template, and date range.
      operationId: getUserWorkflows
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/WorkflowStatusFilter'
        - $ref: '#/components/parameters/WorkflowTemplateFilter'
        - $ref: '#/components/parameters/CreatedAfterFilter'
        - $ref: '#/components/parameters/CreatedBeforeFilter'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
      responses:
        '200':
          description: Workflows retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowListResponse'
        '400':
          description: Bad request - invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Workflows]
      summary: Create new workflow
      description: |
        Create a new workflow based on a template with provided context.
        The workflow will be created in 'draft' status and can be executed later.
      operationId: createWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
            examples:
              simpleWorkflow:
                summary: Simple workflow creation
                value:
                  template: "data-analysis"
                  context:
                    title: "Sales Data Analysis"
                    description: "Analyze Q4 sales data"
                    dataSource: "sales_db"
                  settings:
                    priority: "normal"
                    notifyOnComplete: true
      responses:
        '201':
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /workflows/templates:
    get:
      tags: [Workflows]
      summary: Get workflow templates
      description: |
        Retrieve available workflow templates. Authentication is optional but
        authenticated users may see additional templates based on their role.
      operationId: getWorkflowTemplates
      security: [] # Optional authentication
      parameters:
        - $ref: '#/components/parameters/TemplateCategoryFilter'
        - $ref: '#/components/parameters/TemplateComplexityFilter'
      responses:
        '200':
          description: Templates retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowTemplatesResponse'
  /workflows/{workflowId}:
    get:
      tags: [Workflows]
      summary: Get workflow details
      description: Retrieve detailed information about a specific workflow.
      operationId: getWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: Workflow retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [Workflows]
      summary: Update workflow
      description: Update workflow settings, context, or metadata.
      operationId: updateWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkflowRequest'
      responses:
        '200':
          description: Workflow updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Workflows]
      summary: Delete workflow
      description: |
        Delete a workflow. Only workflows in 'draft', 'completed', or 'failed'
        status can be deleted. Running workflows must be stopped first.
      operationId: deleteWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: Workflow deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteWorkflowResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied or workflow cannot be deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from /auth/login or /auth/refresh endpoints.
        Include in Authorization header: "Bearer <token>"

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  parameters:
...
    Page:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    Limit:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    WorkflowStatusFilter:
      name: status
      in: query
      description: Filter workflows by status
      schema:
        type: string
        enum: [draft, running, waiting_for_human, completed, failed, stopped]
      example: "running"
    WorkflowTemplateFilter:
      name: template
      in: query
      description: Filter workflows by template
      schema:
        type: string
        example: "data-analysis"
    CreatedAfterFilter:
      name: createdAfter
      in: query
      description: Filter workflows created after this date
      schema:
        type: string
        format: date-time
      example: "2024-01-01T00:00:00.000Z"
    CreatedBeforeFilter:
      name: createdBefore
      in: query
      description: Filter workflows created before this date
      schema:
        type: string
        format: date-time
      example: "2024-12-31T23:59:59.999Z"
    SortBy:
      name: sortBy
      in: query
      description: Field to sort by
      schema:
        type: string
        enum: [createdAt, updatedAt, startedAt, completedAt, title, status]
        default: createdAt
      example: "createdAt"
    SortOrder:
      name: sortOrder
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      example: "desc"
    TemplateCategoryFilter:
      name: category
      in: query
      description: Filter templates by category
      schema:
        type: string
        enum: [analytics, decision, content, admin, automation]
        example: "analytics"
...
  schemas:
...
    CreateWorkflowRequest:
      type: object
      required: [template, context]
      properties:
        template:
          type: string
          description: Workflow template identifier
          example: "data-analysis"
        context:
          type: object
          description: Workflow context data
          properties:
            title:
              type: string
              description: Workflow title
              example: "Sales Data Analysis"
            description:
              type: string
              description: Workflow description
              example: "Analyze Q4 sales data and generate insights"
            dataSource:
              type: string
              description: Data source identifier
              example: "sales_db"
            parameters:
              type: object
              description: Additional parameters
              additionalProperties: true
          required: [title, description]
        settings:
          type: object
          description: Workflow settings
          properties:
            priority:
              type: string
              enum: [low, normal, high, urgent]
              default: normal
              description: Workflow execution priority
            notifyOnComplete:
              type: boolean
              default: true
              description: Send notification on completion
            timeoutMinutes:
              type: integer
              minimum: 1
              maximum: 1440
              default: 60
              description: Maximum execution time in minutes
            retryAttempts:
              type: integer
              minimum: 0
              maximum: 5
              default: 0
              description: Number of retry attempts on failure
          additionalProperties: false
      additionalProperties: false
...
    WorkflowListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Workflows retrieved successfully"
        data:
          type: object
          properties:
            workflows:
              type: array
              items:
                $ref: '#/components/schemas/Workflow'
            pagination:
              $ref: '#/components/schemas/Pagination'
          required: [workflows, pagination]
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
      required: [success, message, data, metadata]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Human-readable error message
          example: "Request validation failed"
        code:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
        details:
          oneOf:
            - type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: Field with error
                    example: "email"
                  message:
                    type: string
                    description: Error message
                    example: "Valid email is required"
                  code:
                    type: string
                    description: Error code for field
                    example: "INVALID_EMAIL"
                required: [field, message]
            - type: object
              additionalProperties: true
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-01T00:00:00.000Z"
        requestId:
          type: string
          format: uuid
          description: Request identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        path:
          type: string
          description: Request path
          example: "/api/v1/auth/login"
        retryAfter:
          type: integer
          description: Seconds to wait before retrying
          example: 60
      required: [success, message, code, timestamp]

  # Reusable Request Bodies
...